Move capture_serverless into aws.py
